Rename test function, make 'bad' files binary & add another test
