+++ tests/circuitpython/bad4.toml
@@ -1,0 +1,1 @@
+string="
